+++ src/linodes/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import IndexPage from '~/linodes/layouts/IndexPage';
+import ListLinodesPage from '~/linodes/layouts/ListLinodes';
 import LinodeIndex from './linode/index';
 import NotFound from 'linode-components/dist/errors/NotFound';
 
@@ -10,7 +10,7 @@ const LinodesIndex = (props) => {
   return (
     <Switch>
       <Route component={LinodeIndex} path={`${path}/:linodeLabel`} />
-      <Route component={IndexPage} exact path={`${path}/`} />
+      <Route component={ListLinodesPage} exact path={`${path}/`} />
       <Route component={NotFound} />
     </Switch>
   );
@@ -23,28 +23,3 @@ LinodesIndex.propTypes = {
 };
 
 export default LinodesIndex;
-
-
-/**
- * /lindoes
- *
- * - /:label (dashboard)d
- *
- * - - /networking (summary)
- * - - - /transfer
- * - - - /sharing
- * - - - /resolvers
- *
- * - - /rebuild (index)
- *
- * - - /resize (index)
- *
- * - - /rescue (index)
- *
- * - - /backups (summary)
- * - - - /settings
- *
- * - - /settings (display)
- * - - - /alerts
- * - - - /advanced
- */
